+++ pkgs/jnigen/example/kotlin_plugin/lib/kotlin_bindings.dart
@@ -36,6 +36,8 @@ import 'dart:core' show Object, String;
 import 'package:jni/_internal.dart' as jni$_;
 import 'package:jni/jni.dart' as jni$_;
 
+const _versionCheck = jni$_.JniVersionCheck(1, 0);
+
 /// from: `Example`
 extension type Example._(jni$_.JObject _$this) implements jni$_.JObject {
   static final _class = jni$_.JClass.forName(r'Example');